34.2.Smarter Loading: AJAX status as State{Try...catch; Pass loading down to product-list}
So instead, above the try, add let response. This simply declares the variable outside of the try...catch scope so that it's available in the entire created function. Now, remove the const from response and then I'll return from the catch. So if we hit the catch, just exit. Finally, move the this.products = response.data code outside of the catch. Now if that line has a problem, it won't be silenced: we'll have to deal with it!
Whether or not you should use the try...catch just depends on your situation. I probably wouldn't do this because, if my API endpoint is failing, I have bigger problems: my site is broken! Giving the user a graceful error is nice, but maybe I'll save that for V2.
However, if you do have a valid situation where an AJAX request might fail - like if you're sending data to the server that might fail validation - then this is how you can catch that error and deal with it. We'll talk about sending data in the next tutorial.

Pass loading down to product-list
Okay, we now have the loading data on our smart catalog component. Let's pass that into the product-list component so that we can use it to hide or show the loading spinner. Split the product-list onto multiple lines and then add :loading="loading".
And now that we're passing the loading prop, in index.vue, update the props so we can receive it: add a new loading prop with type: Boolean and required: true.
We can now simplify the template: we want to show the loading animation if loading is true. And we also want to show these product cards down here, if we are ! loading. This second spot isn't super important, but it doesn't hurt to have it!