34.3.Smarter Loading: AJAX status as State{Try...catch; Pass loading down to product-list}
Adding a "No Products" Message after Loading
Time to check things out! Yep! You can already see that the snacks page no longer has the loading spinner. And my other pages work just fine.
Well... except it would be even better with a "no products found" message! And now, we can easily add that.
After the <loading /> component, add an h5 with a v-show directive. This will hold that "no products found" message... which means that we want it to show if we are not loading but products.length === 0.
If that's our situation, print a helpful message. And... there it is! Our snacks page - except for the fact that there are no snacks - works great.

Adding Loading to the Sidebar
The products loading part is now works flawlessly. But there is one other spot that we're loading with AJAX that does not have any loading info: the categories sidebar!
We're actually going to fix this soon by making the categories load instantly. But since they are still loading via AJAX, let's add the loading component there as well. Open up sidebar.vue: this is the component that makes the AJAX request for the categories and renders them in its template.
To do this right, should we add another loading data like we just did in catalog? We totally could! And that's probably a great option. But... I'm going to cheat because I know that my app will never have zero categories. If that ever happened, it would probably mean I accidentally emptied my database. Yikes!
Instead, I am going to use the categories.length to figure out if we're loading. But to be extra organized, let's do this via a computed property called loading. Inside return this.categories.length === 0.
If there are no categories, then we are loading! The nice thing about using a computed property is that it will let us use a simple loading variable in the template. And later, if we did want to change this to data, that would be super easy.
Ok: to use this in the template, first import the loading component: import Loading from '@/components/loading'. Then add the components key with Loading inside.
Finally, up in the template, right after the h5, we'll say <loading with v-show="loading".
I love it!
And when we move over to the browser... I'm hoping to see the loading animation right before the categories load. That was super quick! But it was there. We have proper loading on both sides!
Next, I want to start organizing our AJAX calls: we currently make them from inside of sidebar.vue and catalog.vue. That's maybe ok, but I'd like to explore a better way to organize these.