35.2. AJAX Services{Service: Return the Full Response? Data?}
Service: Return the Full Response? Data?
Before we keep going, I want to mention a design decision that I made. In a service like products-service, we are returning the Promise from Axios. That means that when we use await or chain a normal .then(), what we will ultimately receive is the response... which we can then use to say things like response.data.
But if you want, you could go a bit further in the service and add .then((response) => response.data).
By doing this, our function still returns a Promise. But instead of the payload of the promise being the full AJAX response, it will be the JSON data. To make the code work in catalog, we would set the function directly to the data variable.
That makes the function a bit nicer to use. But... I'm going to completely undo all of that. Why? Changing the function to directly give you the JSON data is nice. The problem is if you ever needed to read something from the response, like a header. If we only returned the data, reading a header wouldn't be possible.