35.3. AJAX Services{Service: Return the Full Response? Data?}
That's why I typically keep my functions simple and return the Promise that resolves to the entire Response. I need to do more work in my component, but I also have more power.
Refresh one more time to make sure nothing broke! All good.
Next, hmm... I don't mind some of the loading on this page, but the categories in the sidebar are starting to get on my nerves! It makes the page look incomplete while it loads because the categories feel like part of the initial page structure. We can fix that by passing the categories from the server directly into Vue!