37.1. Faking AJAX calls: Reading Synchronously{Whoops! We're Returning Different Dat}
The categories data is now available as a global variable: window.categories.
In sidebar.vue, we're calling fetchCategories(), which lives in our fancy new categories-service module. You can find this at assets/js/services/categories-service.js.
To switch this from an AJAX call to the global variable, just return window.categories. Celebrate by removing the axios import on top.
This is what I like about centralizing the fetchCategories() method: we don't need to run around and change our code in a bunch of places: just here. Well... that's not quite true yet. One problem is that our function does not return a Promise anymore! It now returns an array. That could affect code that uses this.

Whoops! We're Returning Different Data
But... let's ignore that for a moment! Move over and refresh. Bah! It's broken:

  Cannot read property hydra:member of undefined

We did change the function from returning a Promise to returning an array. But that's actually not what broke our code! The real problem is that, down in created(), fetchCategories() no longer gives us a response object with a data key!
Let's back up: it's actually ok to use await on a non-async function: JavaScript just grabs the return value from the function and gives it to us: no error. The real problem is that the value that the previous code gave us was a response object... but now we're returning an array of categories. This means that response.data is undefined!