37.2. Faking AJAX calls: Reading Synchronously{Whoops! We're Returning Different Dat}
The simplest way to fix this is right here: comment out the old this.categories line and directly say this.categories = fetchCategories(). We don't need the await anymore, but it doesn't hurt anything.
Simple enough! Back on the browser, it's already showing the categories on the left. And if we reload... yes! The categories are instantly there! Woo! There is still a slight delay before the styles load, but that will only happen in dev mode: we did that to allow hot module replacement. So, mission accomplished!