37.3. Faking AJAX calls: Reading Synchronously{Whoops! We're Returning Different Dat}
Returning a Promise from the Sync Method
One of the cool things about isolating our AJAX calls into functions like fetchCategories() is that, when we call that function, we don't need to know or care if it's talking to an API, or which API, or if we're just loading the data locally from local storage or a variable.
But... that's not really true right now. Because when we changed this from using the API to the global variable, we changed what this function returned! We changed it from returning a Promise to an array and we also changed the actual data that's returned from a response object with data and hydra:member properties to an array. This meant that, once this function was synchronous, we needed to update any code that called this.
And... that's fine. If you know that you're going to load categories synchronously via window.categories, then you can just update your code to reflect that. It's not a huge problem.
But when we changed to the global variable, we could have also written our function in a way that did not break any code that used it: we could have returned a Promise.
Let's try that: return new Promise(). This needs one argument: a callback with resolve and reject. If you've never seen a Promise like this, check out our ES6 tutorial all about them. They're fascinating.
Anyways, once our work is done - which will be instantly since we don't need to make any AJAX calls, call resolve() and pass it the data that we want the promise to return. Now, we're not really making an AJAX request... so we can't exactly return the same data as before because... we don't have a response. But we know that what we really care about is that this Promise returns a object with a data key and a hydra:member key below it. Let's at least fake that here: add data set to an object and hydra:member set to window.categories. I'll remove the extra return at the bottom and, above the function, once again, advertise that we return a Promise!
Now, back in sidebar.vue, we can revert all of our changes and use the exact code we had before: const response = await fetchCategories().
And... that's it! The Promise is a little funny... because it will always resolve immediately. But that's fine! When it resolves, it will return data that's not exactly the same as an Axios response, but it's close.
Moment of truth! At our browser... it looks like it's working. Let's refresh. Yes! It definitely works. So this is a great way to get dynamic data from your server without needing an AJAX call when you really want something to be available almost instantly.
Next, let's use the categories data and currentCategoryId to print the name of the category on each page. To do that, we'll need to make sure all of that data lives in the right component.