38.1. Passing Props vs Fetching Directly{Creating the Title Component; Watch out for Protected Element Names}
At this point, we have access to which category we're currently on and all the category data. Let's use all that goodness to print a better title: we can now print the category name when we're on a category page.
Check out catalog.vue: this is where the h1 lives. Ok, all we need to do is use the currentCategoryId to find that category in the categories data, and then print its name.
To do this, we could create a computed property right in catalog with all of this logic... and then print it right here. But instead, I'm going to isolate this title area into its own component. Again, when you should move something into its own component is subjective, but I'm planning to reuse this title area in the next tutorial.

Creating the Title Component
In the components/ directory, create a new title.vue file and start the same way as always: with the template. Set this to a <div> and immediately add :class="$style.component". Inside, put the h1 with, for now, a hardcoded "Products".
Next, add the <script> tag with the minimum needed: export default and name: 'Title'.
Finally, at the bottom, because we're already referencing $style, add <style lang="scss" module>. We only need one thing: .component {}, an h1 {} inside, and font-size: 1.7rem.
Perfecto! A nice, simple component to render the title.

Watch out for Protected Element Names
In catalog.vue, let's use this! But when I import it, I'm going to call it TitleComponent from @/components/title. The reason is that, when we add TitleComponent to components, it really means TitleComponent: TitleComponent.
Anyways, the key - TitleComponent in this case - determines the HTML element that we can use in the template. TitleComponent means that we can say <title-component />.
That's great. But if we had called it Title, then, in the template, we would need to use <title />. Do you see the problem? <title> is a real HTML tag! And so, instead of rendering our title component, Vue would just... render a title tag.
In general, if you want to avoid collisions, a best practice - which I admit we have not been following - is to include a - every time you render a component. There's a W3C spec that recommends that custom component always have a dash.
Anyways, when we use <title-component>... I think it's working! I still see "Products".