40.1. The Formidable v-model{Creating the search-bar Component}
Ya know what just occurred to me? We haven't talked at all about form elements yet! And what a wonderful coincidence! Because our next challenge is to add a search bar input to filter the product list.
The search bar will, of course, contain some HTML. It will also need to manage the value of the search bar and help us know when we should filter the product list. That's enough that I think we should isolate this in a new component.

Creating the search-bar Component
In components/, create a new file called search-bar.vue. Add the <template> with <div> and an <input> with class="form-control", a placeholder and type="search".
So nothing special. If you're wondering why I added the div, it's just because we're going to have more than just the input later.
At the bottom add the <script> section with the basic export default and name: 'SearchBar'.
Love it! Over in catalog.vue, let's see... change the div around the title to col-3, and then, below, add a new <div class="col-9">. Inside, we haven't imported the search-bar component yet... but ah! Let's try to use it anyway! Type <sea and hit tab to auto-complete that.
When we did that, because PhpStorm is awesome, it added the import and put this down in the components section. PhpStorm, did we just become best friends?
Let's check the browser. Boom! That's a sweet search bar. Now let's bring it to life