40.2. The Formidable v-model{Binding a data to the Input}
Binding a data to the Input
To filter the product list, we need to know the value of the input. There are a few ways to get this, but the simplest is to add a piece of data to the component that we keep "in sync" with the text of the input. Add data and return an object with one item: searchTerm set to an empty string to start.
To "bind" this data to the input, we need to do two things. First, set the input value to the data: :value="searchTerm".
Now with just that, if move over to the browser and look at the Vue dev tools, we can click on SearchBar, change the searchTerm data and... voilà! The input text updates... which shouldn't be too surprising. That's Vue goodness in action.
But what we can't do yet is type in the box and have it update that data. That's the second part of binding an input to data.