40.3. The Formidable v-model{@input for when the Input Changes}
@input for when the Input Changes
How can we do that? It's lovely: by listening to an event on the input.
Remember: the way we listen to an event is with v-on: and then the name of the event, like click or keydown. Well, in practice, we use the shortcut @ syntax, so @click or @keydown.
In this case, use @input. The input event is a native, normal JavaScript event that's similar to keyup: it will trigger any time the input's value changes.
Inside the quotes, so far, we've set this to something like someMethod. Then, we've added a methods option, with a someMethod function, and put whatever code we want to run right there.
That's totally valid, but if what you need to do is simple, you can also write an expression right inside the attribute. In this case, when the input changes, we only need to set the searchTerm data to the new string.