40.4. The Formidable v-model{Inline v-on Expression}
Inline v-on Expression
To do that, say searchTerm - which we know really means this.searchTerm - equals $event.target.value.
That... deserves an explanation. If we set this to a method name, then the method will receive an event argument. Then we can say event.target.value to get this input's value.
When you write an inline expression, Vue magically makes the event object available as a $event variable.
But... hmmm... ESLint is mad: it doesn't like how I ordered my attributes. From a technical standpoint, we can put these in whatever order we want. But as a standard, ESLint likes to have listeners - like @input - on the bottom.
Ok, let's take our input for a test drive! Back at the browser, find SearchBar on the dev tools. Now, as we type... yea! The data is updating!
What we effectively just did is took a piece of data - searchTerm - and bound it to a form input. When the data changes, the input changes. When the input changes, the data changes.