40.5. The Formidable v-model{Hello v-model! }
Hello v-model!
This is a very common thing to do in Vue. In fact, it's so common that Vue created a special directive just for it.
Check this out: we can delete the @input and :value lines and replace them with one line that will do the exact same thing. It's v-model="searchTerm".
This is one of the last important directives that we haven't already talked about. v-model="searchTerm" literally means: set the value attribute to searchTerm and, on input, update the searchTerm data with the input value. It's identical to what we had before.
Now v-model does act a little bit different with things like checkboxes or select elements, but those are minor normalizations to make v-model work how you want it to work for those elements. Back at the browser, everything is hooked up just like before.
You're probably going to see v-model all the time. For me, it helps to remember what it's really doing behind the scenes: setting the input's value to the searchTerm data and, on change, updating that data for you.
Next: let's use this search term to filter our product list as the user is typing!